-Added a menu to the store
-Added products to the store

+++ Assets/Scripts/Coins/Coin.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class Coin : MonoBehaviour
 {
-    [SerializeField] private int _money;
+    [SerializeField] private float _money;
 
     private void OnTriggerEnter2D(Collider2D collision)
     {
